Teaching Only Version/roadmap_swimlane_app/requirements.txt
streamlit>=1.35
pandas>=2.2
openpyxl>=3.1
pydantic>=2.6
matplotlib>=3.9
python-pptx>=1.0.2
pytest>=8.0
